Creates PokeApiCard component
