+++ src/components/organisms/PokeApiCard.styled.ts
@@ -1,0 +1,3 @@
+import styled from "styled-components";
+
+export const StyledPokeApiCard = styled.span``;
