+++ src/components/pages/App.styled.ts
@@ -21,12 +21,6 @@ export const StyledContent = styled.main`
 	margin: auto;
 `;
 
-export const StyledStatus = styled.span`
-	justify-self: center;
-	align-self: center;
-	text-align: center;
-`;
-
 export const StyledGrid = styled.section`
 	grid-template-columns: repeat(auto-fill, minmax(320px, 1fr));
 	gap: ${({ theme }) => theme.spaces.l};
